# api_clients/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DefaultSelection_Prisma._36_SchoolPayload_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"name",
					"id"
				],
				"type": "object"
			},
			"School": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_SchoolPayload_",
				"description": "Model School"
			},
			"SchoolSearchResult": {
				"properties": {
					"pageInfo": {
						"properties": {
							"total": {
								"type": "integer",
								"format": "int32"
							},
							"endCursor": {
								"type": "string",
								"nullable": true
							},
							"hasNextPage": {
								"type": "boolean"
							}
						},
						"required": [
							"total",
							"endCursor",
							"hasNextPage"
						],
						"type": "object"
					},
					"edges": {
						"items": {
							"properties": {
								"node": {
									"$ref": "#/components/schemas/School"
								},
								"cursor": {
									"type": "string"
								}
							},
							"required": [
								"node",
								"cursor"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"pageInfo",
					"edges"
				],
				"type": "object"
			},
			"SchoolErrorType": {
				"enum": [
					"SCHOOL_NOT_FOUND",
					"SCHOOL_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"Pick_SchoolError.message-or-type_": {
				"properties": {
					"message": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"properties": {},
								"additionalProperties": {
									"type": "string"
								},
								"type": "object"
							}
						]
					},
					"type": {
						"$ref": "#/components/schemas/SchoolErrorType"
					}
				},
				"required": [
					"message",
					"type"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"SchoolErrorType.SCHOOL_ALREADY_EXISTS": {
				"enum": [
					"SCHOOL_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"SchoolAlreadyExistsError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_SchoolError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/SchoolErrorType.SCHOOL_ALREADY_EXISTS"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"Pick_School.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewSchool": {
				"$ref": "#/components/schemas/Pick_School.name_"
			},
			"SchoolErrorType.SCHOOL_NOT_FOUND": {
				"enum": [
					"SCHOOL_NOT_FOUND"
				],
				"type": "string"
			},
			"SchoolNotFoundError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_SchoolError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/SchoolErrorType.SCHOOL_NOT_FOUND"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"UpdatedSchool": {
				"$ref": "#/components/schemas/Pick_School.name_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/schools/search": {
			"get": {
				"operationId": "SearchSchools",
				"responses": {
					"200": {
						"description": "Schools Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SchoolSearchResult"
								}
							}
						}
					}
				},
				"tags": [
					"Schools"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/schools/create": {
			"post": {
				"operationId": "CreateSchool",
				"responses": {
					"201": {
						"description": "School Created Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/School"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"name": "Example School"
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "School already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SchoolAlreadyExistsError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Schools"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewSchool"
							}
						}
					}
				}
			}
		},
		"/schools/{id}": {
			"get": {
				"operationId": "GetSchool",
				"responses": {
					"200": {
						"description": "School Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/School"
								}
							}
						}
					},
					"404": {
						"description": "School not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SchoolNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Schools"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSchool",
				"responses": {
					"200": {
						"description": "School Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/School"
								}
							}
						}
					},
					"404": {
						"description": "School not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SchoolNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Schools"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatedSchool"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSchool",
				"responses": {
					"200": {
						"description": "School Deleted Successfully"
					},
					"404": {
						"description": "School not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SchoolNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Schools"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}